TakeTestController.java: replaced findAll() by findById(), implemented findById()

+++ src/main/java/eu/programit/repository/QuestionRepository.java
@@ -6,7 +6,9 @@ import org.springframework.stereotype.Component;
 import eu.programit.domain.Question;
 
 @Component
-public interface QuestionRepository extends CrudRepository<Question, Long> { 
+public interface QuestionRepository extends CrudRepository<Question, Integer> {
 	
+	//Question findOne(int questionId);
+
 }
